requirements.txt
speechrecognition
pyttsx3
whisper
sounddevice
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
streamlit
langdetect